+++ src/validation.cpp
@@ -746,13 +746,15 @@ AcceptToMemoryPoolWorker(const Config &config, CTxMemPool &pool,
 
         // Calculate in-mempool ancestors, up to a limit.
         CTxMemPool::setEntries setAncestors;
-        size_t nLimitAncestors =
-            gArgs.GetArg("-limitancestorcount", DEFAULT_ANCESTOR_LIMIT);
+        size_t nLimitAncestors = gArgs.GetArg(
+            "-limitancestorcount",
+            GetDefaultAncestorLimit(consensusParams, chainActive.Tip()));
         size_t nLimitAncestorSize =
             gArgs.GetArg("-limitancestorsize", DEFAULT_ANCESTOR_SIZE_LIMIT) *
             1000;
-        size_t nLimitDescendants =
-            gArgs.GetArg("-limitdescendantcount", DEFAULT_DESCENDANT_LIMIT);
+        size_t nLimitDescendants = gArgs.GetArg(
+            "-limitdescendantcount",
+            GetDefaultDescendantLimit(consensusParams, chainActive.Tip()));
         size_t nLimitDescendantSize =
             gArgs.GetArg("-limitdescendantsize",
                          DEFAULT_DESCENDANT_SIZE_LIMIT) *